add feature for insert&delete item in various list types

+++ app.js
@@ -17,6 +17,7 @@ app.use(express.static("public"));
 // ejs- embedded javascript
 app.set("view engine", "ejs");
 let itemName = "";
+let listName;
 /* ---------------------------------- Database ---------------------------------- */
 const dbUrl = process.env.DB_URL; // get the url from .env file
 mongoose.set("strictQuery", false);
@@ -46,26 +47,13 @@ const listCategorySchema = new mongoose.Schema({
 const Todo = mongoose.model("todolist", todoSchema);
 
 const ListCategory = mongoose.model("todoListCategory", listCategorySchema);
-
-function insertDefaultItems() {
-  const list1 = new Todo({
-    item: "Welcome to your todolist",
-  });
-
-  const list2 = new Todo({
-    item: "Hit the + button to add a new item",
-  });
-
-  const defaultItems = [list1, list2];
-
-  Todo.insertMany(defaultItems, function (err) {
-    if (err) {
-      console.log(err);
-    } else {
-      console.log("Successfully saved default items to database");
-    }
-  });
-}
+const list1 = new Todo({
+  item: "Welcome to your todolist",
+});
+const list2 = new Todo({
+  item: "Hit the + button to add a new item",
+});
+const defaultItems = [list1, list2];
 
 /* ---------------------------------- Database ---------------------------------- */
 
@@ -75,18 +63,25 @@ const workItems = [];
 
 app.get("/", function (req, res) {
   const today = myDate.day();
-  // console.log(today);
   // gettting data from database
   let itemArray = [];
-
+  listName = today;
   Todo.find({}, function (err, results) {
     if (err) {
       console.log(err);
       res.send("Error in fetching data from database");
     } else {
-      // data is fetched from database
+      // data is fetched from database.
       if (results.length === 0) {
-        insertDefaultItems();
+        // no list in the collections. Inserting default list
+        Todo.insertMany(defaultItems, function (err) {
+          if (err) {
+            console.log(err);
+          } else {
+            console.log("Successfully saved default items to database");
+          }
+        });
+
         res.redirect("/");
         return;
       }
@@ -101,63 +96,83 @@ app.post("/", function (req, res) {
   //takes input from frontend and store it in the array
   // items.push(req.body.newItem);
   itemName = req.body.newItem;
-  // saving data to database
+  // console.log(listName);
   const item = new Todo({
     item: itemName,
   });
+  if (listName === myDate.day()) {
+    // its default list
 
-  item.save();
+    // saving root/main list to database
+    item.save();
+    //displays the updated todo list
+    res.redirect("/");
+  } else {
+    // another todolist
 
-  //displays the updated todo list
-  res.redirect("/");
+    // find the corresponding list category and insert the item in it
+    ListCategory.findOne({ listCategory: listName }, (err, foundList) => {
+      foundList.items.push(item);
+      foundList.save((err, result) => {
+        if (err) console.error("Error while saving the list: ", err);
+        else {
+          // redirect user to the corresponding page to view the updated list.
+          res.redirect("/" + listName);
+        }
+      });
+    });
+  }
 });
 
 // to delete list
 app.post("/delete", (req, res) => {
   const deletedItemId = req.body.itemId;
+  const listName = req.body.listName;
 
-  Todo.deleteOne({ _id: deletedItemId }, (err) => {
-    if (err) console.error(err);
-    // else console.log("deleted!");
-    res.redirect("/");
-  });
+  if (listName === myDate.day()) {
+    // default list
+    Todo.deleteOne({ _id: deletedItemId }, (err) => {
+      if (err) console.error(err);
+      // else console.log("deleted!");
+      res.redirect("/");
+    });
+  } else {
+    // find the list category, then find the item name, and remove it.
+    ListCategory.findOneAndUpdate(
+      { listCategory: listName },
+      { $pull: { items: { _id: deletedItemId } } },
+      (err) => {
+        if (err) console.error(err);
+        else {
+          // deleted
+          res.redirect("/" + listName);
+        }
+      }
+    );
+  }
 });
-
-// app.get("/work", function (req, res) {
-//   res.render("list", { kindOfList: "Work List", itemArray: workItems });
-// });
 
 // auto create route and provide list category feature to the user
 app.get("/:listCategory", (req, res) => {
   // console.log(req.params.category);
   let itemArray = [];
   const listCategory = req.params.listCategory;
-
+  listName = listCategory;
   // check whether the list category exists or not
   ListCategory.findOne({ listCategory: listCategory }, (err, results) => {
-
     if (results) {
       // if category is found in collection
-      console.log("list already exist");
-    }
-    else {
-      // category not exist yet
+      res.render("list", { kindOfList: listName, itemArray: results.items });
+    } else {
+      // category not exist yet, creating new category
       const list = new ListCategory({
         listCategory: listCategory,
-        items: [],
+        items: defaultItems,
       });
-
       list.save();
+      res.render("list", { kindOfList: listName, itemArray: itemArray });
     }
   });
-
-  res.render("list", { kindOfList: listCategory, itemArray: itemArray });
-});
-
-app.post("/work", function (req, res) {
-  workItems.push(req.body.newItem);
-  // console.log("work array updated");
-  res.redirect("/work");
 });
 
 app.listen(port, function () {
